+++ gridstatus/tests/base_test_iso.py
@@ -329,12 +329,22 @@ class BaseTestISO:
         assert is_numeric_dtype(df["Load"])
 
     def _check_forecast(self, df):
+        # support both old and new format
+        # eventually all ISOs should return publish time
         assert set(df.columns) == set(
             [
                 "Time",
                 "Interval Start",
                 "Interval End",
                 "Forecast Time",
+                "Load Forecast",
+            ],
+        ) or set(df.columns) == set(
+            [
+                "Time",
+                "Interval Start",
+                "Interval End",
+                "Publish Time",
                 "Load Forecast",
             ],
         )
@@ -347,24 +357,26 @@ class BaseTestISO:
             series,
         ) | pd.core.dtypes.common.is_timedelta64_ns_dtype(series)
 
+    lmp_cols = [
+        "Time",
+        "Interval Start",
+        "Interval End",
+        "Market",
+        "Location",
+        "Location Type",
+        "LMP",
+        "Energy",
+        "Congestion",
+        "Loss",
+    ]
+
     def _check_lmp_columns(self, df, market):
         # todo in future all ISO should return same columns
         # maybe with the exception of "LMP" breakdown
         self._check_time_columns(df)
 
         assert set(
-            [
-                "Time",
-                "Interval Start",
-                "Interval End",
-                "Market",
-                "Location",
-                "Location Type",
-                "LMP",
-                "Energy",
-                "Congestion",
-                "Loss",
-            ],
+            self.lmp_cols,
         ).issubset(df.columns)
 
         assert len(df["Market"].unique()) == 1